feat: extract runtime.js
